Added LimitReducer
